fix(rust): review fixes metrics3a
Apply 4 spec-driven fixes:
1. RoomMetrics.reconnect_count: i32 → u64
2. RoomMetrics.rejected: HashMap<String, i32> → HashMap<String, u64>
3. MetricPercentiles p50/p95: remove skip_serializing_if (keep explicit null)
4. MetricsReport.value: i32 → f64; MetricsHealthSnapshot counters → u64

Wire-shape parity: manager receives explicit null fields, no truncation risk.

Gate: GO ✅ | cargo test -p razzoozle-server: 19 passed

+++ rust/protocol/src/player.rs
@@ -167,7 +167,7 @@ pub enum MetricKind {
 #[serde(rename_all = "camelCase")]
 pub struct MetricsReport {
     pub kind: MetricKind,
-    pub value: i32,
+    pub value: f64,
 }
 
 /// Metrics subscription (client->server).
@@ -186,10 +186,8 @@ pub struct MetricsSubscribe {
 #[ts(export)]
 #[serde(rename_all = "camelCase")]
 pub struct MetricPercentiles {
-    #[serde(skip_serializing_if = "Option::is_none")]
     #[ts(optional)]
     pub p50: Option<i32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
     #[ts(optional)]
     pub p95: Option<i32>,
     pub count: i32,
@@ -204,8 +202,8 @@ pub struct MetricsHealthSnapshot {
     pub rtt: MetricPercentiles,
     pub clock_offset: MetricPercentiles,
     pub answer_ack: MetricPercentiles,
-    pub reconnect_count: i32,
-    pub rejected: std::collections::HashMap<String, i32>,
+    pub reconnect_count: u64,
+    pub rejected: std::collections::HashMap<String, u64>,
 }
 
 // ============================================================================
